refactor(cli): split build_cmd into concern modules
The build command module mixed orchestration with venv/env setup,
lifecycle phases, per-service injectors, and project persistence; those
now live in build_environment, build_lifecycle, build_injectors, and
build_persistence, with build_cmd reduced to the Click command and thin
orchestration. All helpers are re-exported so import paths are unchanged.
